Began reorganizing code into multiple .cc files for my own sanity; verified previous functionality is preserved
