+++ subsystems/intt/InttMonDraw.h
@@ -16,6 +16,7 @@
 #include <TSystem.h>
 
 #include <TLine.h>
+#include <TPolyLine.h>
 #include <TText.h>
 
 #include <TH1D.h>
@@ -34,6 +35,7 @@
 class InttMonDraw : public OnlMonDraw
 {
  public:
+  // InttMonDraw.cc
   InttMonDraw(std::string const&);
   ~InttMonDraw() override;
 
@@ -54,40 +56,66 @@ class InttMonDraw : public OnlMonDraw
   typedef std::map<std::string, struct Option_s> Options_t;
   static Options_t OPTIONS;
 
-  // ...
+  // Intt_o_Helpers.cc
+  struct Margin_s
+  {
+    double t = 0.1;
+    double b = 0.1;
+    double l = 0.1;
+    double r = 0.1;
+  };
+  void static DrawPad(TPad*, TPad*, struct Margin_s const& = (struct Margin_s){
+    .t = 0.1, .b = 0.1, .l = 0.1, .r = 0.1});
+  void static CdPad(TPad*);
+
+  Color_t static GetFeeColor(int const&);
+
+  // Intt_o_FelixBcoFphxBco.cc
+  struct FelixBcoFphxBco_s
+  {
+	double cnvs_width, cnvs_height;
+	double disp_frac, lgnd_frac;
+	double disp_text_size;
+	double lgnd_box_width, lgnd_box_height, lgnd_text_size;
+	std::string name;
+  } static const m_FelixBcoFphxBco;
+  void static DrawFelixBcoFphxBco(std::string const&);
+  void static DrawFelixBcoFphxBco_DispPad();
+  void static DrawFelixBcoFphxBco_LgndPad();
+  void static DrawFelixBcoFphxBco_SubPads();
+  void static DrawFelixBcoFphxBco_SubPad(int);
+
+  // In Intt_o_Whatever.cc
+  // To be removed
+  static void DrawBcoDiff(std::string const&);
   static void DrawHitsVsEvt(std::string const&);
 
-  // BCO histograms
-  static void DrawBcoDiff(std::string const&);
-
-  // GlobalChip-Channel idiom
   static void GlobalChipLocalChannelHead(std::string const&);
   static void DrawGlobalChipMap(std::string const&);
   static void InttGlobalChipExec(const std::string&, int);
   static void DrawLocalChannelMap(std::string const&, struct INTT::Indexes_s);
   static void InttLocalChannelExec(const std::string&, int, int, int, int);
 
-  // GlobalLadder-Chip idiom
   static void GlobalLadderLocalChipHead(std::string const&);
   static void DrawGlobalLadderMap(std::string const&);
   static void InttGlobalLadderExec(const std::string&, int);
   static void DrawLocalChipMap(std::string const&, struct INTT::Indexes_s);
   static void InttLocalChipExec(const std::string&, int, int, int);
 
-  // GlobalChip-Channel methods
   static void PrepGlobalChipHists_Hitmap(std::string const&, TH2D**);
   static void PrepLocalChannelHists_Hitmap(std::string const&, TH2D**, struct INTT::Indexes_s&);
   static void PrepGlobalChipHists_NLL(std::string const&, TH2D**);
 
-  // GlobalLadder-Chip methods
   static void PrepGlobalLadderHists_Interface(std::string const&, TH2D**);
   static void PrepLocalChipHists_Hitmap(std::string const&, TH2D**, struct INTT::Indexes_s&);
 
  private:
+  // InttMonDraw.cc
   int MakeCanvas(const std::string& name);
   int DrawServerStats();
   TCanvas* TC[1] = {nullptr};
   TPad* transparent[1] = {nullptr};
+
   //===  Constants for Drawing  ===//
   static constexpr double NUM_SIG = 2.0;
 
@@ -111,10 +139,6 @@ class InttMonDraw : public OnlMonDraw
   static constexpr double Y_LABEL_TEXT_SIZE = 0.5;
   static constexpr double X_LABEL_TEXT_SIZE = 0.5;
   //===  ~Constants for Drawing  ===//
-
-  //===  Drawing Methods   ===//
-  static void DrawPad(TPad*, TPad*);
-  //===  ~Drawing Methods  ===//
 };
 
 #endif
